refactor: move fuzz tests for profile.rs into their own file
