+++ profiling/src/internal/profile/mod.rs
@@ -1,6 +1,9 @@
 // Copyright 2021-Present Datadog, Inc. https://www.datadoghq.com/
 // SPDX-License-Identifier: Apache-2.0
 
+#[cfg(test)]
+mod fuzz_tests;
+
 use self::api::UpscalingInfo;
 use super::*;
 use crate::api;
@@ -544,408 +547,7 @@ impl Profile {
 #[cfg(test)]
 mod api_tests {
     use super::*;
-    use bolero::TypeGenerator;
-    use std::collections::{HashMap, HashSet};
-
-    fn assert_sample_types_eq(
-        profile: &pprof::Profile,
-        expected_sample_types: &[owned_types::ValueType],
-    ) {
-        assert_eq!(
-            profile.sample_types.len(),
-            expected_sample_types.len(),
-            "Sample types length mismatch"
-        );
-        for (typ, expected_typ) in profile
-            .sample_types
-            .iter()
-            .zip(expected_sample_types.iter())
-        {
-            assert_eq!(
-                *profile.string_table[typ.r#type as usize],
-                *expected_typ.r#typ
-            );
-            assert_eq!(*profile.string_table[typ.unit as usize], *expected_typ.unit);
-        }
-    }
-
-    fn assert_samples_eq(
-        profile: &pprof::Profile,
-        samples_with_timestamps: &[&owned_types::Sample],
-        samples_without_timestamps: &HashMap<
-            (&Vec<owned_types::Location>, &Vec<owned_types::Label>),
-            Vec<i64>,
-        >,
-        endpoint_mappings: &FxIndexMap<u64, &String>,
-    ) {
-        assert_eq!(
-            profile.samples.len(),
-            samples_with_timestamps.len() + samples_without_timestamps.len(),
-            "Samples length mismatch"
-        );
-
-        let mut expected_timestamped_samples = samples_with_timestamps.iter();
-
-        for sample in profile.samples.iter() {
-            // Recreate owned_locations from vector of pprof::Location
-            let mut owned_locations = Vec::new();
-            for loc_id in sample.location_ids.iter() {
-                let location = &profile.locations[*loc_id as usize - 1];
-                let mapping = &profile.mappings[location.mapping_id as usize - 1];
-                assert!(location.lines.len() == 1);
-                let line = location.lines[0];
-                let function = profile.functions[line.function_id as usize - 1];
-                assert!(!location.is_folded);
-
-                let owned_location = owned_types::Location {
-                    mapping: owned_types::Mapping {
-                        memory_start: mapping.memory_start,
-                        memory_limit: mapping.memory_limit,
-                        file_offset: mapping.file_offset,
-                        filename: profile.string_table[mapping.filename as usize]
-                            .clone()
-                            .into_boxed_str(),
-                        build_id: profile.string_table[mapping.build_id as usize]
-                            .clone()
-                            .into_boxed_str(),
-                    },
-                    function: owned_types::Function {
-                        name: profile.string_table[function.name as usize]
-                            .clone()
-                            .into_boxed_str(),
-                        system_name: profile.string_table[function.system_name as usize]
-                            .clone()
-                            .into_boxed_str(),
-                        filename: profile.string_table[function.filename as usize]
-                            .clone()
-                            .into_boxed_str(),
-                        start_line: function.start_line,
-                    },
-                    address: location.address,
-                    line: line.line,
-                };
-
-                owned_locations.push(owned_location);
-            }
-
-            // Recreate owned_labels from vector of pprof::Label
-            let mut owned_labels = Vec::new();
-            for label in sample.labels.iter() {
-                let key = profile.string_table[label.key as usize]
-                    .clone()
-                    .into_boxed_str();
-
-                if *key == *"end_timestamp_ns" {
-                    // TODO: Check end timestamp label
-                    continue;
-                } else if *key == *"trace endpoint" {
-                    let actual_str = &profile.string_table[label.str as usize];
-
-                    let prev_label: &owned_types::Label = owned_labels
-                        .last()
-                        .expect("Previous label to exist for endpoint label");
-                    let num = prev_label.num as u64;
-                    let expected_str = endpoint_mappings
-                        .get(&num)
-                        .expect("Endpoint mapping to exist");
-
-                    assert_eq!(actual_str, *expected_str);
-                    continue;
-                }
-
-                if label.str != 0 {
-                    let str = Box::from(profile.string_table[label.str as usize].as_str());
-                    owned_labels.push(owned_types::Label {
-                        key,
-                        str: Some(str),
-                        num: 0,
-                        num_unit: None,
-                    });
-                } else {
-                    let num = label.num;
-                    let num_unit = if label.num_unit != 0 {
-                        Some(
-                            profile.string_table[label.num_unit as usize]
-                                .clone()
-                                .into_boxed_str(),
-                        )
-                    } else {
-                        None
-                    };
-                    owned_labels.push(owned_types::Label {
-                        key,
-                        str: None,
-                        num,
-                        num_unit,
-                    });
-                }
-            }
-
-            if let Some(expected_sample) = expected_timestamped_samples.next() {
-                assert_eq!(owned_locations, expected_sample.locations);
-                assert_eq!(sample.values, expected_sample.values);
-                assert_eq!(owned_labels, expected_sample.labels);
-            } else {
-                let key = (&owned_locations, &owned_labels);
-                let expected_values = samples_without_timestamps
-                    .get(&key)
-                    .expect("Value not found for an aggregated sample");
-                assert_eq!(&sample.values, expected_values);
-            }
-        }
-    }
-
-    fn add_sample<'a>(
-        timestamp: &Option<Timestamp>,
-        sample: &'a owned_types::Sample,
-        expected_sample_types: &[owned_types::ValueType],
-        profile: &mut Profile,
-        samples_with_timestamps: &mut Vec<&'a owned_types::Sample>,
-        samples_without_timestamps: &mut HashMap<
-            (&'a Vec<owned_types::Location>, &'a Vec<owned_types::Label>),
-            Vec<i64>,
-        >,
-    ) {
-        let r = profile.add_sample(sample.into(), *timestamp);
-        if expected_sample_types.len() == sample.values.len() {
-            assert!(r.is_ok());
-            if timestamp.is_some() {
-                samples_with_timestamps.push(sample);
-            } else if let Some(existing_values) =
-                samples_without_timestamps.get_mut(&(&sample.locations, &sample.labels))
-            {
-                existing_values
-                    .iter_mut()
-                    .zip(sample.values.iter())
-                    .for_each(|(a, b)| *a = a.saturating_add(*b));
-            } else {
-                samples_without_timestamps
-                    .insert((&sample.locations, &sample.labels), sample.values.clone());
-            }
-        } else {
-            assert!(r.is_err());
-        }
-    }
-
-    /// Fuzzes adding a bunch of samples to the profile.
-    #[test]
-    fn fuzz_add_sample() {
-        bolero::check!()
-            .with_generator((
-                Vec::<owned_types::ValueType>::gen(),
-                Vec::<(Option<Timestamp>, owned_types::Sample)>::gen(),
-            ))
-            .for_each(|(expected_sample_types, samples)| {
-                let sample_types: Vec<_> = expected_sample_types
-                    .iter()
-                    .map(api::ValueType::from)
-                    .collect();
-                let mut expected_profile = Profile::new(SystemTime::now(), &sample_types, None);
-                let mut samples_with_timestamps = Vec::new();
-                let mut samples_without_timestamps: HashMap<
-                    (&Vec<owned_types::Location>, &Vec<owned_types::Label>),
-                    Vec<i64>,
-                > = HashMap::new();
-                for (timestamp, sample) in samples {
-                    add_sample(
-                        timestamp,
-                        sample,
-                        expected_sample_types,
-                        &mut expected_profile,
-                        &mut samples_with_timestamps,
-                        &mut samples_without_timestamps,
-                    );
-                }
-                let profile = pprof::roundtrip_to_pprof(expected_profile).unwrap();
-                assert_sample_types_eq(&profile, expected_sample_types);
-                assert_samples_eq(
-                    &profile,
-                    &samples_with_timestamps,
-                    &samples_without_timestamps,
-                    &FxIndexMap::default(),
-                );
-            })
-    }
-
-    #[test]
-    fn fuzz_add_sample_with_fixed_sample_length() {
-        let sample_length_gen = if cfg!(miri) { 1..=8usize } else { 1..=64usize };
-
-        bolero::check!()
-            .with_generator(sample_length_gen)
-            .and_then(|sample_len| {
-                let sample_types = Vec::<owned_types::ValueType>::gen().with().len(sample_len);
-
-                let timestamps = Option::<Timestamp>::gen();
-                let locations = Vec::<owned_types::Location>::gen();
-                let values = Vec::<i64>::gen().with().len(sample_len);
-                // Generate labels with unique keys
-                let labels = HashSet::<owned_types::Label>::gen();
-
-                let samples = Vec::<(
-                    Option<Timestamp>,
-                    Vec<owned_types::Location>,
-                    Vec<i64>,
-                    HashSet<owned_types::Label>,
-                )>::gen()
-                .with()
-                .values((timestamps, locations, values, labels));
-                (sample_types, samples)
-            })
-            .for_each(|(sample_types, samples)| {
-                let api_sample_types: Vec<_> =
-                    sample_types.iter().map(api::ValueType::from).collect();
-                let mut profile = Profile::new(SystemTime::now(), &api_sample_types, None);
-                let mut samples_with_timestamps = Vec::new();
-                let mut samples_without_timestamps: HashMap<
-                    (&Vec<owned_types::Location>, &Vec<owned_types::Label>),
-                    Vec<i64>,
-                > = HashMap::new();
-
-                let samples: Vec<(&Option<Timestamp>, owned_types::Sample)> = samples
-                    .iter()
-                    .map(|(timestamp, locations, values, labels)| {
-                        (
-                            timestamp,
-                            owned_types::Sample {
-                                locations: locations.clone(),
-                                values: values.clone(),
-                                labels: labels
-                                    .clone()
-                                    .into_iter()
-                                    .collect::<Vec<owned_types::Label>>(),
-                            },
-                        )
-                    })
-                    .collect();
-
-                for (timestamp, sample) in samples.iter() {
-                    profile
-                        .add_sample(sample.into(), **timestamp)
-                        .expect("Failed to add sample");
-                    if timestamp.is_some() {
-                        samples_with_timestamps.push(sample);
-                    } else if let Some(existing_values) =
-                        samples_without_timestamps.get_mut(&(&sample.locations, &sample.labels))
-                    {
-                        existing_values
-                            .iter_mut()
-                            .zip(sample.values.iter())
-                            .for_each(|(a, b)| *a = a.saturating_add(*b));
-                    } else {
-                        samples_without_timestamps
-                            .insert((&sample.locations, &sample.labels), sample.values.clone());
-                    }
-                }
-                let serialized_profile =
-                    pprof::roundtrip_to_pprof(profile).expect("Failed to roundtrip to pprof");
-
-                assert_sample_types_eq(&serialized_profile, sample_types);
-                assert_samples_eq(
-                    &serialized_profile,
-                    &samples_with_timestamps,
-                    &samples_without_timestamps,
-                    &FxIndexMap::default(),
-                );
-            });
-    }
-
-    #[test]
-    fn fuzz_add_endpoint() {
-        bolero::check!()
-            .with_type::<Vec<(u64, String)>>()
-            .for_each(|endpoints| {
-                let mut profile = Profile::new(SystemTime::now(), &[], None);
-                for (local_root_span_id, endpoint) in endpoints {
-                    profile
-                        .add_endpoint(*local_root_span_id, endpoint.into())
-                        .expect("add_endpoint to succeed");
-                }
-                pprof::roundtrip_to_pprof(profile).expect("roundtrip_to_pprof to succeed");
-            });
-    }
-
-    #[test]
-    fn fuzz_add_endpoint_count() {
-        bolero::check!()
-            .with_type::<Vec<(String, i64)>>()
-            .for_each(|endpoint_counts| {
-                let mut profile = Profile::new(SystemTime::now(), &[], None);
-                for (endpoint, count) in endpoint_counts {
-                    profile
-                        .add_endpoint_count(endpoint.into(), *count)
-                        .expect("add_endpoint_count to succeed");
-                }
-                pprof::roundtrip_to_pprof(profile).expect("roundtrip_to_pprof to succeed");
-            });
-    }
-
-    #[derive(Debug, TypeGenerator)]
-    enum Function {
-        AddSample(Option<Timestamp>, owned_types::Sample),
-        AddEndpoint(u64, String),
-    }
-
-    #[derive(Debug, TypeGenerator)]
-    struct ApiFunctionCalls {
-        sample_types: Vec<owned_types::ValueType>,
-        operations: Vec<Function>,
-    }
-
-    #[test]
-    fn fuzz_api_function_calls() {
-        let sample_length_gen = if cfg!(miri) { 1..=8usize } else { 1..=64usize };
-
-        bolero::check!()
-            .with_generator(sample_length_gen)
-            .and_then(|sample_len| {
-                let sample_types = Vec::<owned_types::ValueType>::gen().with().len(sample_len);
-                let operations = Vec::<Function>::gen();
-
-                (sample_types, operations)
-            })
-            .for_each(|(sample_types, operations)| {
-                let api_sample_types: Vec<_> =
-                    sample_types.iter().map(api::ValueType::from).collect();
-                let mut profile = Profile::new(SystemTime::now(), &api_sample_types, None);
-                let mut samples_with_timestamps: Vec<&owned_types::Sample> = Vec::new();
-                let mut samples_without_timestamps: HashMap<
-                    (&Vec<owned_types::Location>, &Vec<owned_types::Label>),
-                    Vec<i64>,
-                > = HashMap::new();
-                let mut endpoint_mappings: FxIndexMap<u64, &String> = FxIndexMap::default();
-
-                for operation in operations {
-                    match operation {
-                        Function::AddSample(timestamp, sample) => {
-                            add_sample(
-                                timestamp,
-                                sample,
-                                sample_types,
-                                &mut profile,
-                                &mut samples_with_timestamps,
-                                &mut samples_without_timestamps,
-                            );
-                        }
-                        Function::AddEndpoint(local_root_span_id, endpoint) => {
-                            profile
-                                .add_endpoint(*local_root_span_id, endpoint.into())
-                                .expect("add_endpoint to succeed");
-                            endpoint_mappings.insert(*local_root_span_id, endpoint);
-                        }
-                    }
-                }
-
-                let pprof_profile = pprof::roundtrip_to_pprof(profile).unwrap();
-                assert_sample_types_eq(&pprof_profile, sample_types);
-                assert_samples_eq(
-                    &pprof_profile,
-                    &samples_with_timestamps,
-                    &samples_without_timestamps,
-                    &endpoint_mappings,
-                );
-            })
-    }
+    use std::collections::HashMap;
 
     #[test]
     fn interning() {
